Task 0 betty update

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -3,13 +3,13 @@
 /**
  * print_list - a function that prints all the elements of a list_t list.
  *
- * @h: a const list of type list_t.
+ * @h: a struct pointer of type list_t.
  * Return: number of nodes.
  */
 
 size_t print_list(const list_t *h)
 {
-	int counter = 0;
+	size_t counter = 0;
 
 	while (h != NULL)
 	{
